added postgres, dockerized offers-ng

+++ offersproject/offers/models.py
@@ -48,7 +48,7 @@ class Category(models.Model):
 class Offer(models.Model):
     title = models.CharField(max_length=200)
     description = models.TextField()
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=100, decimal_places=2)
     created_at = models.DateTimeField(auto_now_add=True)
     category = models.ForeignKey(Category, null=True, on_delete=models.SET_NULL)
